pyproject.toml
[tool.poetry]
name = "ktv_repeat_monitor"
version = "0.1.0"
description = ""
authors = ["tsukachu <concon.ch.1543@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fire = "^0.4.0"
SQLAlchemy = "^1.4.18"
psycopg2 = "^2.9.1"
alembic = "^1.6.5"
simple-settings = "^1.0.0"
SQLAlchemy-Utils = "^0.37.7"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = {extras = ["colors"], version = "^5.8.0"}
flake8 = "^3.9.2"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
